some bugs and issues fix

+++ main.py
@@ -11,7 +11,8 @@ import os
 global queue, i, video_title_name, playlist_added_msg, first_added_of_pl, pl_urls, track_id
 global play_flag, dwnld_pl_flag
 
-music_dir = 'D:/ITAN/'
+music_dir = 'D:/ITAN/Download/'
+vibe_dir = 'D:/ITAN/вайб/'
 
 old_files = os.listdir(music_dir)
 for fil in old_files:
@@ -32,8 +33,10 @@ async def play_song(ctx, url):
         return
     else:
         channel = ctx.message.author.voice.channel
-    try: await channel.connect()
-    except Exception as err: print(err)
+    try:
+        await channel.connect()
+    except Exception as err:
+        print(err)
 
     try:
         queue = queue
@@ -56,7 +59,7 @@ async def play_song(ctx, url):
 
     except Exception as err:
         print('>>download err', err)
-        await ctx.send(f"Somenthing went wrong - {err}")
+        await ctx.send(f"Something went wrong - {err}")
 
 
 async def start_playing(voice_channel, ctx):
@@ -64,12 +67,21 @@ async def start_playing(voice_channel, ctx):
     while True:
         if play_flag:
             try:
-                voice_channel.play(discord.FFmpegPCMAudio(executable="ffmpeg.exe", source=music_dir + queue[i]))
-                await ctx.send(f'**Пою: **`{queue[i][2:]}`'.replace('.webm', '').replace('_', ' '))
-                try:
-                    os.remove(music_dir + queue[i-1])
-                    queue[i-1] = ''
-                except: pass
+                voice_channel.play(discord.FFmpegPCMAudio(executable="ffmpeg.exe", source=queue[i]))
+                track_title = f'{queue[i]}'.split(sep='/')[-1].split(sep='.')[0]
+                track_title = track_title[2:] if track_title[0].isnumeric() else track_title
+                await ctx.send(f'**Пою: **`{track_title}`'.replace('.webm', '').replace('_', ' '))
+
+                await asyncio.sleep(1)
+                if queue[i].split(sep='/')[-2] == 'Download':
+                    if i > 0:
+                        try:
+                            os.remove(music_dir + queue[i-1])
+                            queue[i-1] = ''
+                        except:
+                            pass
+                    else:
+                        pass
                 i += 1
             except:
                 if dwnld_pl_flag:
@@ -77,9 +89,10 @@ async def start_playing(voice_channel, ctx):
                     await asyncio.sleep(1)
                     global pl_urls
                     await download_playlist(pl_urls)
-                else: pass
+                else:
+                    pass
 
-            await asyncio.sleep(3)
+            await asyncio.sleep(4)
 
         elif not play_flag:
             break
@@ -94,15 +107,17 @@ async def download(url, ctx):
         pl_urls = []
         for url in playlist_urls:
             pl_urls.append(url)
-        print(pl_urls)
         first_of_pl = [pl_urls[0], f'{track_id} ' + get_video_title(pl_urls[0])]
-        YoutubeDL(set_video_name(first_of_pl[1])).download(pl_urls[0])
+        try:
+            YoutubeDL(set_video_name(first_of_pl[1])).download(pl_urls[0])
+        except:
+            YoutubeDL(set_video_name(f'{track_id}_N_{track_id}')).download(pl_urls[0])
         track_id += 1
 
         files = os.listdir(music_dir)
         for n in range(len(files)):
-            if files[n] not in queue:
-                queue.append(files[n])
+            if (music_dir + files[n]) not in queue:
+                queue.append(music_dir + files[n])
                 first_added_of_pl = files[n][2:]
             else:
                 pass
@@ -119,21 +134,13 @@ async def download(url, ctx):
         track_id += 1
         files = os.listdir(music_dir)
         for n in range(len(files)):
-            if files[n] not in queue:
-                queue.append(files[n])
+            if (music_dir + files[n]) not in queue:
+                queue.append(music_dir + files[n])
                 added = files[n][2:]
                 dwnld_msg_text = f'**Добавлен: **`{added}`'.replace('.webm', '').replace('_', ' ')
             else:
                 pass
         await ctx.send(dwnld_msg_text)
-
-    old_files = os.listdir(music_dir)
-    for fil in old_files:
-        if fil not in queue:
-            try:
-                os.remove(music_dir + fil)
-            except:
-                pass
 
 
 async def download_playlist(urls):
@@ -148,34 +155,36 @@ async def download_playlist(urls):
 
             files = os.listdir(music_dir)
             for n in range(len(files)):
-                if files[n] not in queue:
-                    queue.append(files[n])
+                if (music_dir + files[n]) not in queue:
+                    queue.append(music_dir + files[n])
                     added.append(files[n][2:])
                 else:
                     pass
             if added:
                 dots = '.' * (len(urls) - len(added) - 1)
-                text = f'**Плейлист добавлен**\n```-{first_added_of_pl}\n{added}\n{dots}```'
-                text = text.replace("', '", '\n-')
-                text = text.replace("['", '-').replace("']", '')
-                text = text.replace('- ', '-').replace('.webm', '')
-                await playlist_added_msg.edit(content=text)
-        except Exception as err: print('1 ', err)
-    print('pl dwnlded')
+                pl_append_text = f'**Плейлист добавлен**\n```-{first_added_of_pl}\n{added}\n{dots}```'
+                pl_append_text = pl_append_text.replace("', '", '\n-')
+                pl_append_text = pl_append_text.replace("['", '-').replace("']", '')
+                pl_append_text = pl_append_text.replace('- ', '-').replace('.webm', '')
+                await playlist_added_msg.edit(content=pl_append_text)
+        except Exception as err:
+            print('>>dwnld_playlist ', err)
+    print('pl downloaded')
 
 
 def get_video_title(url):
     response = requests.get(url)
     soup = BeautifulSoup(response.text, 'html.parser')
     title = soup.find('title').text
-    title = title.replace('- YouTube', '')
+    title = title.replace('- YouTube', '').replace('/', '!').replace('|', 'I').replace('<', '%').replace('>', '%')
+    title = title.replace(':', ';').replace('*', '^').replace('?', 'a').replace('"', '`').replace("\\", '!').replace('.', ' ')
     return title
 
 
 def set_video_name(video_name):
     ydl_opts = {
         'format': 'bestaudio/best',
-        'outtmpl': f'{music_dir}{video_name.title()}.%(ext)s',
+        'outtmpl': f'{music_dir}{video_name}.%(ext)s',
         'restrictfilenames': True,
         'nocheckcertificate': True,
         'ignoreerrors': True,
@@ -237,15 +246,82 @@ async def clear_queue(ctx):
 @bot.command(name='очередь', help='Посмотреть что дальше')
 async def print_queue(ctx):
     global queue, i
+    queue_next = []
     if len(queue[i:]) > 0:
-        queue_msg = f'**Дальше по списку: **```{queue[i:]}```'
-        queue_msg = queue_msg.replace("', '", '\n-')
+        for n in queue[i:]:
+            queue_next.append(n.split('/')[-1])
+        queue_msg = f'**Дальше по списку: **```{queue_next}```'
+        queue_msg = queue_msg.replace("', '", '\n-').replace('.mp3', '')
         queue_msg = queue_msg.replace("['", '-').replace("']", '')
         queue_msg = queue_msg.replace('_', ' ').replace('.webm', '')
         await ctx.send(queue_msg)
     else:
         queue_msg = '**В очереди нет треков**'
         await ctx.send(queue_msg)
+
+
+@bot.command(name='+вайб', help='+вайб')
+async def play_vibe(ctx):
+    import random
+    global queue, i, track_id, dwnld_pl_flag, play_flag
+
+    if not ctx.message.author.voice:
+        await ctx.send("{} is not connected to a voice channel".format(ctx.message.author.name))
+        return
+    else:
+        channel = ctx.message.author.voice.channel
+    try:
+        await channel.connect()
+    except Exception as err:
+        print(err)
+
+    vibe_list = os.listdir(vibe_dir)
+    voice_client = ctx.message.guild.voice_client
+    try:
+        if queue[-1].split('/')[-1] not in vibe_list:
+            queue[i:] = []
+            try:
+                voice_client.stop()
+            except:
+                pass
+    except NameError:
+        queue = []
+        track_id = i = 0
+        dwnld_pl_flag = False
+    except Exception as err:
+        print('>>+vibe not name err ', err)
+
+    random.shuffle(vibe_list)
+    dots = '.' * len(vibe_list)
+    vibe_msg_text = f'**Добавлен +вайб: **```\n{dots}```'
+    vibe_added_msg = await ctx.send(vibe_msg_text)
+
+    added = []
+    for vibe in vibe_list:
+        if vibe not in queue:
+            queue.append(vibe_dir + vibe)
+            added.append(vibe)
+        else:
+            pass
+
+        if added:
+            dots = '.' * (len(vibe_list) - len(added))
+            pl_append_text = f'**Плейлист добавлен**\n```{added}\n{dots}```'
+            pl_append_text = pl_append_text.replace("', '", '\n-')
+            pl_append_text = pl_append_text.replace("['", '-').replace("']", '')
+            pl_append_text = pl_append_text.replace('- ', '-').replace('.mp3', '')
+            await vibe_added_msg.edit(content=pl_append_text)
+
+    old_files = os.listdir(music_dir)
+    for fil in old_files:
+        os.remove(music_dir + fil)
+
+    play_flag = True
+    try:
+        await start_playing(voice_client, ctx)
+    except Exception as err:
+        print('>>play err', err)
+
 
 dotenv.load_dotenv('C:/Users/tortm/PyProjects/ITAN_MusicianBot/.env')
 TOKEN = os.getenv('TOKEN')
